Some minor cleanup. Fixing several tests that had dependencies on the year, or otherwise started failing recently. Also moving ScriptRow to be in the same package as the other Script* classes for better code organization.
