+++ src/test/java/gov/noaa/pfel/erddap/dataset/EDDGridFromDapTests.java
@@ -11137,6 +11137,7 @@ class EDDGridFromDapTests {
    * @throws Throwable if trouble
    */
   @org.junit.jupiter.api.Test
+  @TagThredds // external server is failing to respond, so disable the test for now
   void testActualRange2() throws Throwable {
     // String2.log("\n*** EDDGridFromDap.testActualRange2");
     int language = 0;
